add agent UI
