+++ src/code/dialogs/teamMembershipList.js
@@ -3,6 +3,8 @@ import Sortable from "../sortable";
 import WasabeeTeam from "../model/team";
 import wX from "../wX";
 
+import AgentUI from "../ui/agent";
+
 const TeamMembershipList = WDialog.extend({
   statics: {
     TYPE: "teamMembershipList",
@@ -54,7 +56,9 @@ const TeamMembershipList = WDialog.extend({
         value: (agent) => agent.name,
         sort: (a, b) => a.localeCompare(b),
         format: async (cell, value, agent) =>
-          cell.appendChild(await agent.formatDisplay(this.options.teamID)),
+          cell.appendChild(
+            await AgentUI.formatDisplay(agent, this.options.teamID)
+          ),
       },
       {
         name: wX("SQUAD"),
